file wrappers and history versioned
added in file wrappers from the rust-kms repo since they are fairly
generic. they currently require that the type in question implement
serde de/serialize but that may change in the future.

also moved over the versioned struct from rust-kms since it was also
generic. added in a non-synchronized version for good measure.